Refactor student into modulith design
