+++ src/main/java/vv/pms/allocation/internal/ProjectAllocationRepository.java
@@ -5,5 +5,7 @@ import vv.pms.allocation.ProjectAllocation;
 import java.util.Optional;
 
 public interface ProjectAllocationRepository extends JpaRepository<ProjectAllocation, Long> {
+
+    // Custom finder to check if a project is already allocated
     Optional<ProjectAllocation> findByProjectId(Long projectId);
 }
