+++ src/main/java/vv/pms/professor/ProfessorService.java
@@ -13,18 +13,12 @@ public class ProfessorService {
 
     private final ProfessorRepository repository;
 
-    /**
-     * Constructor for ProfessorService
-     */
     public ProfessorService(ProfessorRepository repository) {
         this.repository = repository;
     }
 
     /**
-     * Adds a new professor to the system.
-     * @param name  The name of the professor
-     * @param email The email of the professor
-     * @return The created Professor entity
+     * CREATE: Adds a new professor to the system.
      */
     public Professor addProfessor(String name, String email) {
         if (repository.findByEmail(email).isPresent()) {
@@ -35,10 +29,8 @@ public class ProfessorService {
     }
 
     /**
-     * Retrieves a professor by their ID.
+     * READ: Retrieves a professor by their ID.
      * This is crucial for cross-module communication (e.g., by the Allocation module).
-     * @param id The ID of the professor
-     * @return An Optional containing the Professor if found, or empty if not found
      */
     @Transactional(readOnly = true)
     public Optional<Professor> findProfessorById(Long id) {
@@ -46,8 +38,7 @@ public class ProfessorService {
     }
 
     /**
-     * Retrieves all professors.
-     * @return A list of all Professor entities
+     * READ: Retrieves all professors.
      */
     @Transactional(readOnly = true)
     public List<Professor> findAllProfessors() {
@@ -55,11 +46,8 @@ public class ProfessorService {
     }
 
     /**
-     * Removes a professor by ID.
-     * TODO: Consider cascading effects required when deleting a professor
-     * 
-     * @param id The ID of the professor to remove
-     * @throws ProfessorNotFoundException if the professor with the given ID does not exist
+     * DELETE: Removes a professor by ID.
+     * NOTE: Real-world systems require complex checks before deletion (e.g., are they assigned to any projects?).
      */
     public void deleteProfessor(Long id) {
         if (!repository.existsById(id)) {
